chore(mobile): rename account to account-management
