+++ suite-native/app/package.json
@@ -54,8 +54,8 @@
         "@suite-native/config": "workspace:*",
         "@suite-native/fiat-rates": "workspace:*",
         "@suite-native/message-system": "workspace:*",
-        "@suite-native/module-accounts": "workspace:*",
-        "@suite-native/module-accounts-import": "workspace:*",
+        "@suite-native/module-accounts-management": "workspace:*",
+        "@suite-native/module-accounts-management-import": "workspace:*",
         "@suite-native/module-dev-utils": "workspace:*",
         "@suite-native/module-home": "workspace:*",
         "@suite-native/module-onboarding": "workspace:*",
