+++ suite-native/app/src/navigation/AppTabNavigator.tsx
@@ -2,7 +2,7 @@ import { BottomTabBarProps, createBottomTabNavigator } from '@react-navigation/b
 
 import { ReceiveStackNavigator } from '@suite-native/module-receive';
 import { HomeStackNavigator } from '@suite-native/module-home';
-import { AccountsStackNavigator } from '@suite-native/module-accounts';
+import { AccountsStackNavigator } from '@suite-native/module-accounts-management';
 import { SettingsStackNavigator } from '@suite-native/module-settings';
 import { AppTabsParamList, AppTabsRoutes, TabBar } from '@suite-native/navigation';
