+++ suite-native/app/src/navigation/RootStackNavigator.tsx
@@ -2,8 +2,11 @@ import { useSelector } from 'react-redux';
 
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import { AccountDetailScreen, AccountSettingsScreen } from '@suite-native/module-accounts';
-import { AccountsImportStackNavigator } from '@suite-native/module-accounts-import';
+import {
+    AccountDetailScreen,
+    AccountSettingsScreen,
+} from '@suite-native/module-accounts-management';
+import { AccountsImportStackNavigator } from '@suite-native/module-accounts-management-import';
 import {
     RootStackParamList,
     RootStackRoutes,
