+++ suite-native/module-accounts-management/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-native/module-accounts",
+    "name": "@suite-native/module-accounts-management",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
