+++ suite-native/module-accounts-management/src/components/AccountDetailCryptoValue.tsx
