+++ suite-native/module-accounts-management/src/components/AccountDetailGraph.tsx
