+++ suite-native/module-accounts-management/src/components/AccountDetailGraphHeader.tsx
