+++ suite-native/module-accounts-management/src/components/AccountDetailScreenHeader.tsx
