+++ suite-native/module-accounts-management/src/components/AccountSettingsRemoveCoinButton.tsx
