+++ suite-native/module-accounts-management/src/components/CoinPriceCard.tsx
