+++ suite-native/module-accounts-management/src/components/TokenAccountDetailScreenHeader.tsx
