+++ suite-native/module-accounts-management/src/screens/AccountDetailScreen.tsx
