+++ suite-native/module-accounts-management/tsconfig.json
